Fix cicular dependyencies and move joi to common
