+++ packages/common/package.json
@@ -84,6 +84,7 @@
   },
   "dependencies": {
     "@zkdb/permission": "0.1.6",
+    "joi": "^17.13.3",
     "mongodb": "^6.11.0",
     "o1js": "1.9.1"
   }
