+++ packages/common/src/index.ts
@@ -1,18 +1,3 @@
-export * from './schema.js';
-export * from './types/collection.js';
-export * from './types/common.js';
-export * from './types/database.js';
-export * from './types/document.js';
-export * from './types/group.js';
-export * from './types/merkle-tree.js';
-export * from './types/metadata.js';
-export * from './types/ownership.js';
-export * from './types/pagination.js';
-export * from './types/proof.js';
-export * from './types/rollup.js';
-export * from './types/schema.js';
-export * from './types/search.js';
-export * from './types/sequencer.js';
-export * from './types/setting.js';
-export * from './types/transaction.js';
-export * from './types/user.js';
+export * from './types';
+export * from './validation';
+export * from './schema';
