+++ packages/serverless/src/apollo/app/transaction.ts
@@ -3,6 +3,8 @@ import {
   TDatabaseRequest,
   TTransactionRequest,
   TTransactionConfirmRequest,
+  databaseName,
+  transactionType,
 } from '@zkdb/common';
 import GraphQLJSON from 'graphql-type-json';
 import Joi from 'joi';
@@ -13,7 +15,6 @@ import {
 } from '../../domain/use-case/transaction.js';
 import { gql } from '../../helper/common.js';
 import { authorizeWrapper } from '../validation.js';
-import { databaseName, transactionType } from './common.js';
 
 export const typeDefsTransaction = gql`
   #graphql
