+++ packages/storage/src/database/base/basic.ts
@@ -1,9 +1,4 @@
-import {
-  ClientSession,
-  CreateCollectionOptions,
-  Document,
-  WithoutId,
-} from 'mongodb';
+import { ClientSession, CreateCollectionOptions, Document } from 'mongodb';
 import logger from '../../helper/logger.js';
 import { DatabaseEngine } from '../database-engine.js';
